Moved caching from EventAggregator to separate class
